improve: split library routes
