+++ server/routes/library/studios.ts
@@ -1,0 +1,31 @@
+import express from "express";
+import {
+  authenticateToken,
+  requireCacheReady,
+} from "../../middleware/auth.js";
+import {
+  findStudios,
+  findStudiosMinimal,
+  updateStudio,
+} from "../../controllers/library/studios.js";
+import { authenticated } from "../../utils/routeHelpers.js";
+
+const router = express.Router();
+
+// All studio routes require authentication
+router.use(authenticateToken);
+
+// Find studios with filters
+router.post("/studios", requireCacheReady, authenticated(findStudios));
+
+// Minimal data for filter dropdowns
+router.post(
+  "/studios/minimal",
+  requireCacheReady,
+  authenticated(findStudiosMinimal)
+);
+
+// Update studio
+router.put("/studios/:id", authenticated(updateStudio));
+
+export default router;
