+++ server/routes/playlist.ts
@@ -10,6 +10,7 @@ import {
   removeSceneFromPlaylist,
   reorderPlaylist,
 } from "../controllers/playlist.js";
+import { authenticated } from "../utils/routeHelpers.js";
 
 const router = express.Router();
 
@@ -17,27 +18,27 @@ const router = express.Router();
 router.use(authenticateToken);
 
 // Get all user playlists
-router.get("/", getUserPlaylists);
+router.get("/", authenticated(getUserPlaylists));
 
 // Get single playlist with items
-router.get("/:id", getPlaylist);
+router.get("/:id", authenticated(getPlaylist));
 
 // Create new playlist
-router.post("/", createPlaylist);
+router.post("/", authenticated(createPlaylist));
 
 // Update playlist
-router.put("/:id", updatePlaylist);
+router.put("/:id", authenticated(updatePlaylist));
 
 // Delete playlist
-router.delete("/:id", deletePlaylist);
+router.delete("/:id", authenticated(deletePlaylist));
 
 // Add scene to playlist
-router.post("/:id/items", addSceneToPlaylist);
+router.post("/:id/items", authenticated(addSceneToPlaylist));
 
 // Remove scene from playlist
-router.delete("/:id/items/:sceneId", removeSceneFromPlaylist);
+router.delete("/:id/items/:sceneId", authenticated(removeSceneFromPlaylist));
 
 // Reorder playlist items
-router.put("/:id/reorder", reorderPlaylist);
+router.put("/:id/reorder", authenticated(reorderPlaylist));
 
 export default router;
